Fix timestamp size

+++ src/controllers/event.rs
@@ -7,7 +7,7 @@ pub async fn new(pool: &'static Pool, user_id: i64, event: NewEvent, headers: wa
     debug!("{:?}", headers);
     let mac = headers.get("MAC_ADDRESS").ok_or(Error::NothingFound)?.to_str().map_err(|_|Error::BadData)?.to_string();
     let version = headers.get("VERSION").ok_or(Error::BadData)?.to_str().map_err(|_|Error::BadData)?.to_string();
-    let time_running: u16 = headers.get("TIME_RUNNING").ok_or(Error::BadData)?.to_str().map_err(|_|Error::BadData)?.parse().map_err(|_|Error::BadData)?;
+    let time_running: u64 = headers.get("TIME_RUNNING").ok_or(Error::BadData)?.to_str().map_err(|_|Error::BadData)?.parse().map_err(|_|Error::BadData)?;
     let vcc: u16 = headers.get("VCC").ok_or(Error::BadData)?.to_str().map_err(|_|Error::BadData)?.parse().map_err(|_|Error::BadData)?;
     let free_heap: u64 = headers.get("FREE_HEAP").ok_or(Error::BadData)?.to_str().map_err(|_|Error::BadData)?.parse().map_err(|_|Error::BadData)?;
     let free_stack: u32 = headers.get("FREE_STACK").ok_or(Error::BadData)?.to_str().map_err(|_|Error::BadData)?.parse().map_err(|_|Error::BadData)?;
